improved test and clone external

+++ src/kdtree_interface_cppPosit.cpp
@@ -15,13 +15,13 @@ kdd_factory_t cppPosit_factories[] = {
 	
 	{"posit8_0f", FACTORYPOSIT(int8_t, 8, 1, float, false) },
 	{"posit16_1f", FACTORYPOSIT(int16_t, 16, 1, float, false) },
-	{"posit32_2f", FACTORYPOSIT(int32_t, 32, 2, float, false) },
+	{"posit32_2f", FACTORYPOSIT(int32_t, 32, 2, double, false) },
 
 	// po     sit
 	//{"posit16_2", FACTORYPOSIT(int16_t, 16, POSIT16_ES, uint32_t, false) },
 	{"posit8_2", FACTORYPOSIT(int8_t, 8, POSIT8_ES, uint32_t, false) },
 	{"posit12_2", FACTORYPOSIT(int16_t, 12, POSIT12_ES, uint32_t, false) },
-	
+
 	// tabulated
 	{"posit8tab", FACTORY(posit8) } ,
 	{"posit12tab", FACTORY(posit12) } ,
